+++ packages/plugins/local-storage/tests/dir-utils.test.ts
@@ -45,7 +45,6 @@ describe('searchOnFolders', () => {
       {
         name: 'pkg1',
       },
-
       {
         name: 'pkg2',
       },