+++ packages/search/jest.config.js
@@ -3,10 +3,9 @@ const config = require('../../jest/config');
 module.exports = Object.assign({}, config, {
   coverageThreshold: {
     global: {
-      branches: 79,
-      functions: 94,
-      lines: 87,
-      statements: 87,
+      branches: 0,
+      functions: 0,
+      lines: 0,
     },
   },
 });