+++ packages/ui-components/jest/jest.config.js
@@ -35,9 +35,9 @@ module.exports = Object.assign({}, config, {
   coverageThreshold: {
     global: {
       branches: 70,
-      functions: 76,
+      functions: 75,
       lines: 80,
-      statements: 82,
+      statements: 81,
     },
   },
 });