+++ packages/ui-components/jest/server-handlers.ts
@@ -31,6 +31,15 @@ export const handlers = [
   rest.get('http://localhost:9000/-/verdaccio/data/sidebar/jquery', (req, res, ctx) => {
     return res(ctx.json(require('./api/jquery-sidebar.json')));
   }),
+  rest.get('http://localhost:9000/-/verdaccio/data/sidebar/JSONStream', (req, res, ctx) => {
+    return res(ctx.status(401));
+  }),
+  rest.get('http://localhost:9000/-/verdaccio/data/sidebar/semver', (req, res, ctx) => {
+    return res(ctx.status(500));
+  }),
+  rest.get('http://localhost:9000/-/verdaccio/data/sidebar/kleur', (req, res, ctx) => {
+    return res(ctx.status(404));
+  }),
   rest.get('http://localhost:9000/-/verdaccio/data/sidebar/glob', (req, res, ctx) => {
     return res(ctx.json(require('./api/glob-sidebar.json')));
   }),