+++ packages/ui-components/src/AppTest/pages/Front/Home.ts
@@ -1,0 +1,3 @@
+import { Home } from '../../../index';
+
+export default Home;