+++ packages/logger/logger-commons/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @verdaccio/logger-commons
 
+## 7.0.0-next-7.13
+
+### Patch Changes
+
+- Updated dependencies [9db1554]
+  - @verdaccio/logger-prettify@7.0.0-next-7.2
+  - @verdaccio/core@7.0.0-next-7.13
+
+## 7.0.0-next-7.12
+
+### Patch Changes
+
+- @verdaccio/core@7.0.0-next-7.12
+
+## 7.0.0-next-7.11
+
+### Patch Changes
+
+- @verdaccio/core@7.0.0-next-7.11
+
 ## 7.0.0-next-7.10
 
 ### Patch Changes