+++ packages/plugins/audit/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+## 12.0.0-next-7.13
+
+### Patch Changes
+
+- Updated dependencies [a99a4bb]
+  - @verdaccio/config@7.0.0-next-7.13
+  - @verdaccio/core@7.0.0-next-7.13
+
+## 12.0.0-next-7.12
+
+### Patch Changes
+
+- @verdaccio/core@7.0.0-next-7.12
+- @verdaccio/config@7.0.0-next-7.12
+
+## 12.0.0-next-7.11
+
+### Patch Changes
+
+- Updated dependencies [c9962fe]
+  - @verdaccio/config@7.0.0-next-7.11
+  - @verdaccio/core@7.0.0-next-7.11
+
 ## 12.0.0-next-7.10
 
 ### Patch Changes