+++ packages/plugins/ui-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @verdaccio/ui-theme
 
+## 7.0.0-next-7.13
+
+### Patch Changes
+
+- ba53d1e: feat: versions filter by semver range
+
+## 7.0.0-next-7.12
+
+## 7.0.0-next-7.11
+
+### Minor Changes
+
+- c9962fe: feat: forbidden user interface
+
 ## 7.0.0-next-7.10
 
 ### Patch Changes