+++ packages/store/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @verdaccio/store
 
+## 7.0.0-next-7.13
+
+### Patch Changes
+
+- Updated dependencies [a99a4bb]
+  - @verdaccio/config@7.0.0-next-7.13
+  - @verdaccio/hooks@7.0.0-next-7.13
+  - @verdaccio/loaders@7.0.0-next-7.13
+  - @verdaccio/local-storage@12.0.0-next-7.13
+  - @verdaccio/proxy@7.0.0-next-7.13
+  - @verdaccio/search@7.0.0-next-7.2
+  - @verdaccio/logger@7.0.0-next-7.13
+  - @verdaccio/core@7.0.0-next-7.13
+  - @verdaccio/tarball@12.0.0-next-7.13
+  - @verdaccio/url@12.0.0-next-7.13
+  - @verdaccio/utils@7.0.0-next-7.13
+
+## 7.0.0-next-7.12
+
+### Patch Changes
+
+- Updated dependencies [cce258e]
+  - @verdaccio/search@7.0.0-next-7.1
+  - @verdaccio/core@7.0.0-next-7.12
+  - @verdaccio/config@7.0.0-next-7.12
+  - @verdaccio/tarball@12.0.0-next-7.12
+  - @verdaccio/url@12.0.0-next-7.12
+  - @verdaccio/hooks@7.0.0-next-7.12
+  - @verdaccio/loaders@7.0.0-next-7.12
+  - @verdaccio/local-storage@12.0.0-next-7.12
+  - @verdaccio/proxy@7.0.0-next-7.12
+  - @verdaccio/utils@7.0.0-next-7.12
+  - @verdaccio/logger@7.0.0-next-7.12
+
+## 7.0.0-next-7.11
+
+### Patch Changes
+
+- Updated dependencies [c9962fe]
+  - @verdaccio/config@7.0.0-next-7.11
+  - @verdaccio/hooks@7.0.0-next-7.11
+  - @verdaccio/loaders@7.0.0-next-7.11
+  - @verdaccio/local-storage@12.0.0-next-7.11
+  - @verdaccio/proxy@7.0.0-next-7.11
+  - @verdaccio/core@7.0.0-next-7.11
+  - @verdaccio/tarball@12.0.0-next-7.11
+  - @verdaccio/url@12.0.0-next-7.11
+  - @verdaccio/utils@7.0.0-next-7.11
+  - @verdaccio/logger@7.0.0-next-7.11
+
 ## 7.0.0-next-7.10
 
 ### Patch Changes